+++ frontend/ws_bridge.py
@@ -48,6 +48,7 @@ class WebSocketBridge:
         'console_out',
         'get_joymaping',
         'get_keymapping',
+        'get_mainmenu_config',
         'set_button_mapping',
         'launch_table',
         'update_frontend_dof_for_table',