+++ RELEASE_NOTES.md
@@ -1,11 +1,10 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v1.1.71
+This is the offical release of vpinfe v1.1.72
 
 ### What's New
-- add hook rom / fix upper/lower case rom bug.
-- add vpinplay calls to vpinfe-core.js
+- Add hiding quit button from mainmenu.  In mangerui under "Settings".
 
 ### Fixes
 None